requirements.txt
flask
flask-cors
mysql-connector-python
Pillow
python-dotenv
Werkzeug
